% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gene}
\alias{gene}
\title{gene}
\format{
An object of class \code{data.frame} with 200 rows and 1 columns.
}
\usage{
gene
}
\description{
Just test number(Description)
}
\examples{
  head(gene)
}
\keyword{datasets}
