% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICMC.r
\name{ICMC}
\alias{ICMC}
\title{Title}
\usage{
ICMC(cancer)
}
\arguments{
\item{cancer}{type of cancer(There are 17 cancer types to choose from : "BLCA","BRCA","CESC","COAD","ESCA","GBM","HNSC","KICH","LAML","LIHC","LUSC","PAAD","PRAD","SKCM","STAD","THCA","UCEC")}
}
\value{
Marker methylation levels correlate with immune cell infiltration
}
\description{
Title
}
\examples{
library(DrMyth)
ICMC("UCEC")
}
