% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICMC_plot.r
\name{ICMC_plot}
\alias{ICMC_plot}
\title{Title}
\usage{
ICMC_plot(gene, cancer)
}
\arguments{
\item{gene}{methylation marker for cancer(The marker gene in the results after ICMC analysis)}

\item{cancer}{type of cancer(There are 17 cancer types to choose from : "BLCA","BRCA","CESC","COAD","ESCA","GBM","HNSC","KICH","LAML","LIHC","LUSC","PAAD","PRAD","SKCM","STAD","THCA","UCEC")}
}
\value{
Marker methylation levels correlate with immune cell infiltration
}
\description{
Title
}
\examples{
library(DrMyth)
ICMC("UCEC")
ICMC_plot("MLH1","UCEC")
ICMC_plot(c("MLH1","WT1"),"UCEC")
}
