% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEA.r
\name{MSEA}
\alias{MSEA}
\title{Main functions}
\usage{
MSEA(datalist, marker)
}
\arguments{
\item{datalist}{list of methylation genes}

\item{marker}{list of methylation marker genes}
}
\value{
methylation enrichment results
}
\description{
The Main function of DrMyth
}
\examples{
library(DrMyth)
data(gene)
data(marker)
MSEA(gene,marker)
}
