% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSEA_plot.r
\name{MSEA_plot}
\alias{MSEA_plot}
\title{Title}
\usage{
MSEA_plot(plotdata)
}
\arguments{
\item{plotdata}{table of methylation enrichment result(table after MSEA enrichment)}
}
\value{
bubble plot of methylation enrichment analysis results
}
\description{
Title
}
\examples{
library(DrMyth)
data(gene)
data(marker)
p <- MSEA(gene,marker)
MSEA_plot(p)
}
